requirements.txt
PyQt5>=5.15.0
pyyaml>=6.0
python-libnmap
python-nmap
reportlab>=3.6.0
Jinja2>=3.0.0
requests>=2.28.0
python-nmap>=0.7.1